Remove unnecessary allocations on JCache#get path
Profiling showed that lambdas were being allocated on every read, but
were not doing anything that useful. Instead those two cases are now
inlined. For a read (common case) the adapter does not create any
garbage (stack allocations only).

The majority of the time was spent on the annotation proxies provided
by the RI. That was due to their lookup overhead and allocations.