Avoid unnecessary lambda allocation on insert
If a new entry is being inserted into a cache and a writer isn't used,
then a `computeIfAbsent` can instead be replaced with a `putIfAbsent`.
This avoids an unnecessary allocation, which adds GC pressure.

This is in response to CASSANDRA-10855 perf run which was dominated by
insertions with a cache miss due to the uniform distribution. This is
at least a little optimization, though there's not much we can do for
that scenario.