AsyncCacheLoader for asynchronous caches (fixes #53)
Creating an AsyncLoadingCache previously required passing a CacheLoader
to Caffine.buildAsync(), optionally as a lambda. This is convenient
when the computations are written in a synchronous fashion and the
cache provides the asynchronous wrapping.

That API was clunky when the computation is defined asynchronously,
such as when bubbling up a future from another library. In that case
the asyncLoad default method had to be overriden in addition to load
(which may never be called). It offered all the power needed but in a
combersome fashion.

AsyncCacheLoader resolves this API flaw by providing the asynchronous
computations that a loader can make (asyncLoad, asyncLoadAll, and
asyncReload). It can be used as a lambda for buildAsync, e.g.
(key, executor) -> future.

Backwards compatibility is retained by having the CacheLoader extend
AsyncCacheLoader. The asyncLoad and asyncLoadAll were already defined
as defaults, so only asyncReload was introduced. This approach is
nice by providing symmetry, low integration effort, and minimal impact
to the API.