Reduce the jar size by 400kb (codegen)
When introducing code generation for the entry and cache, it was overly
optimistic and aggressive. That resulted in a larger binary and slow
byte code verification. The type hierarchy was optimized and the least
useful cases were promoted back to declared references.

The CacheLoader reference is another memory optimization that can be
rolled back. It should be very commonly used, as we strongly recommend
using a self populating cache instead of a manual one. The difference
is 14k lines generated (400kb), which is a lot less byte code that
might be verified.