Add optimized spliterator implementations
The default spliterator did not provide any characteristics, like
distinct or concurrent. These optimized ones are better suited if
for stream processing.

The tests identified a bug in ConcurrentHashMap which incorrectly
estimates the size for any odd map after a split. When checking the
ConcurrentSkipListMap behavior, it reports Integer.MAX_VALUE for any
split of a populated map. This issue has been reported.