Avoid ForkJoinPool wrapper causing GC churn
Submitting a runnable into FJP causes it to be wrapped with an adapter
to a ForkJoinTask. For maximumSize this is less significant due to the
overhead of the maintenance work (CountMinSketch hashing, etc). That's
thousands of objects (still not good), but other cache types create
millions. For weakKeys this halfs the read throughput.

See [1] for more details.

[1] https://github.com/orbit/orbit/pull/144