Initialize the frequency sketch if initial capacity set
The frequency sketch, like the hash table, is lazily sized as entries
are added. This is delayed until the cache is at least 50% full and only
then begins recording frequencies. For the maximum size the total entry
count is known, while for the maximum weight it resizes based on the
map's size.

The 50% full laziness is based on observing that many caches are over
sized to give a safety net bounding. This is especially true when
expirable.

In the case of a maximum weight, many times the user has a rough idea
of the expected entry count (e.g. maximum weight / avg entry byte size).
If they are already sizing the hash table then we should size the other
data structures similarly.