Update memory overhead analysis to show aligned size
Now ignores the executor, which used to be not present by codegen but
that was causing too much byte code bloat. This caused the base size
to include ForkJoinPool.commonPool() which is quite hefty.

The previous numbers on the wiki were for 1.x, so now updated. These
are rough estimates and the insight that can be gleemed is tiny. This
primarily lets us see if there's any suprising differences to
investigate.