Improve correctness to match theoretical W-TinyLFU
When the same random hash seed is used, there was a 0.5% loss on the
multi1 trace. Likely other small traces had a similar difference. This
was due to a sizing difference and early promotion of eden entries into
the main space. That led to prematurely designating the warm-up entries
as protected, causing their eligibility for eviction to be delayed.

These were intentional, but for a conurrency optimization that didn't
pan out. The idea had been that the hottest entries could be detected and
the policy work skipped, thereby avoiding read buffer contention. Since
that trick wasn't beneficial it was never enabled and removed.

This won't effect large traces very much, but is confusing on small ones.
Now the only remaining difference is due to the buffer being lossy when
full, which causes a 0.01% difference. That could be fixed by an if
statement to convey full early instead of late, but it seems unnecessary
given Pareto principle.