Add admission rate to simulator
This probably should have been added a long time ago, but doesn't have
a lot of direct insight. It shows that TinyLFU admits between 2-10% of
the entries, depending on the eviction policy. This follows the general
observation that most arrivals pollute the cache and should be quickly
discarded. Policies like LIRS and SLRU are based on that, and TinyLFU
does so explicitly.

An interesting next step is to determine how to inject a feedback cycle
into TinyLFU to correct for mispredictions. That's probably best done
using an adaptive window size, e.g. by adjusting when a recently
rejected entry would be discarded twice (indicating a sparse burst).