Further proving out Adaptive W-TinyLFU is awesome
The first attempt showed that we should adjust when multiple rejections
are detected. It grew the window in recency traces and kept it small in
frequency traces. But, as a bloom filter it didn't properly adapt when
multiple types were run in sequence. It favored a large window.

The next attempt showed that if we should constrain the growth to when
we also detect another accesses. This quickly increased and decreased
the window based on the changing worklaod. However the quick hack used
TinyLFU's frequency (10x max size sample period). It would be fine for
small caches but the recency signal would be less relevant for larger
ones due to the long sample time.

This version uses only the maximum size sample period for both signals.
This was done by hacking another CountMin Sketch variant, but perhaps
two bloom filters would have been better. Deciding on the right sketch
is a future task, as is making this code less horrible. The result was
a further nudge at improving the hit rate with good adaption. I think
we found a winner!