Improved adaptive W-TinyLFU for recency workloads
Added four traces from cache2k for a Java e-commerce webapp. Two traces
are for an ORM and are extremely recency skewed. I suspect that this is
Hibernate's cache, which is usually flushed regularly to avoid stale
data. In this case LRU is the near optimal policy.

TinyLFU does poorly when recency is dominant. W-TinyLFU helps, but the
window is static and may be too small for the workload. The new adaptive
version is excellent, though. It correctly adjusts to emulate LRU while
still being able to return to a leveraging TinyLFU if the workload
changes. This helps to make the policy more robust for different usages.

Caffeine will under perform in this case until adaption is integrated in
a future iteration.