Finalize adaptive W-TinyLFU algorithm
This is (close to) the final version of this new policy. The knobs are
fully externalized, it uses a smaller membership sketch, and evaluates
based on the number of samples taken. The runs with different traces,
alone or paired with others, indicates good performance and adaptive
behavior. The ideas are much more crystalized.