A custom a Bloom filter for policies in the simulator
This implementation is intended to use in TinyLFU-based caches for the
doorkeeper and adaptive sizing. The primary value is that it uses the
same hash functions as the CountMin4. This will allow computing the
hashes once for use in multiple sketches.

The doorkeeper is a technique to reduce the size of the CountMinSketch
by preventing "one-hit-wonders" from needing to be be stored. This is
described in the paper and implemented in go-tinylfu. It was backlogged
in Caffeine and not simulated, but should be introduced at some point.

The adaptive W-TinyLFU uses the sketch to detect mispredictions. This
increases the hit rate by dynamically tuning towards recency or
frequency depending on the workload.

Amusing story. This is based on my old interview prep (code kata) [1]
that later led to Guava's version. When I joined the team kept arguing
that no efficient implementation was possible (believed a SecureRandom
was needed, with the hashCode as the seed?!?). For some reason Googlers
hate reading Wikipedia and research papers, and were surprised when I
showed them a working prototype (yes, a toy hack as I never thought
anyone would look at it). But alas I was then pushed out of discussions
due to being a Noogler and the result was slow due to poor GC (Hasher)
and modulus operations (rejected my power-of-two and no-GC requests).

[1] https://github.com/ben-manes/concurrentlinkedhashmap/blob/wiki/BloomFilter.md