Adaptive refinements to reduce waste
The window captures recency, but that information quickly becomes
useless after a certain age. If the cache is small then it could play a
dominent role and the window needs to be a larger percentage. However it
never needs to be excessively large in total entry count.

256 appears to be the ideal cutoff. For a large cache we can avoid
wasting time and space by using a fixed size window. For a small cache
we can monitor to adapt based on the workload.

The resulting equation is very understandable for a large cache. If the
entry is being evicted from the window we know it has not been used
recently (256 recency distance). TinyLFU then decides whether to keep
the entry based on its frequency, which is now the best hint at reuse.
Most likely it will not and discarding the entry avoids cache
pollution.

A small cache is only slightly more complicated. We monitor a sample
of 1024 to detect mispredictions, using a bloom filter to retain
history. If detected we grow, else shrink, to the threshold limits.