Add doorkeeper to the simulator for analysis
The doorkeeper is a space saving technique to reduce the number of
counters needed by TinyLFU's frequency sketch. A one-hit wonder would
only be a member in the doorkeeper, not make it into the frequency
sketch, and be discarded by the cache. This reduces the number of
entries making it into the count-min sketch, allowing us to save
space without degrading the hit rate.

This optimization is disabled and requires analysis to find the best
parameters. A cursory look shows a 25% space savings for small caches
(short traces).